requirements.txt
Pillow
numpy
fastembed
